+++ ai-packages/ai-deepseek/src/common/api/api-protocol.ts
@@ -1,0 +1,6 @@
+import { OpenAIAPIOptions } from '@celljs/ai-openai';
+
+export const DeepSeekPIOptions = Symbol('DeepSeekPIOptions');
+
+export interface DeepSeekPIOptions extends OpenAIAPIOptions {
+}