+++ ai-packages/ai-deepseek/src/common/model/deepseek-chat-options.ts
@@ -1,0 +1,10 @@
+import { Constant } from '@celljs/core';
+import { OpenAIChatOptions } from '@celljs/ai-openai';
+
+/**
+ * Options for the DeepSeek Chat API.
+ */
+@Constant(DeepSeekChatOptions, new DeepSeekChatOptions())
+export class DeepSeekChatOptions extends OpenAIChatOptions {
+
+}