+++ ai-packages/ai-deepseek/src/common/model/deepseek-embedding-options.ts
@@ -1,0 +1,9 @@
+import { Constant } from '@celljs/core';
+import { OpenAIEmbeddingOptions } from '@celljs/ai-openai';
+
+/**
+ * DeepSeek Embedding Options.
+ */
+@Constant(DeepSeekEmbeddingOptions, new DeepSeekEmbeddingOptions())
+export class DeepSeekEmbeddingOptions extends OpenAIEmbeddingOptions {
+}